fix getPurchases error preventing handling old purchases
getPurchases has an error (which is in a giant generic try catch) which
prevents handling non-consumed transactions on the store. This is a
major issue -- if there are every non-consumed transactions, the user's
device will query the store (and fail) every ten seconds, forever, AND
the purchases will never be consumed.