add restore function to simulate and android
Does not add functionality - simply adds the missing functions so the
api is consistent across each environment.

NOTE: does not actually do anything in either the simulator or on
android. In the simulator this simply creates a fake function that
calls the callback after waiting 2 seconds. On Android this adds
a billingRestore event that gets sent immediately with failure 'not
implemented on android'.